Are there may be more that 1 get webapi method? Hahah, seems, not

+++ ProjectPracticeWeb/Models/Beverage.cs
@@ -3,7 +3,7 @@
     public class Beverage
     {
         public string Name { get; set; }
-        public float Cost { get; set; }
+        public int Cost { get; set; }
         public int Count { get; set; }
     }
 }
